+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 ----------
+Unreleased
+----------
+Changed
+-------
+- Upgraded python-telegram-bot to version 20
+
+----------
 `0.7.0`_ - 2023-01-14
 ----------
 Fixed